requirements.txt
beautifulsoup4==4.10.0
colorama==0.4.4
lxml==4.8.0
Pillow==9.0.1
pytesseract==0.3.9
